requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
tensorflow==2.16.1
numpy==1.26.4
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.4.0
python-dotenv==1.0.0
requests==2.31.0